+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.4.0 (unreleased)
+## 0.3.11 (November 4, 2013)
 
 FEATURES:
 
@@ -9,6 +9,7 @@ IMPROVEMENTS:
 
 * core: builders can now give warnings during validation. warnings won't
   fail the build but may hint at potential future problems.
+* builder/digitalocean: Can now specify a droplet name
 * builder/virtualbox: Can now disable guest addition download entirely
   by setting "guest_additions_mode" to "disable" [GH-580]
 * builder/virtualbox,vmware: ISO urls can now be https [GH-587]
@@ -21,6 +22,9 @@ BUG FIXES:
 * builder/amazon/all: Properly scrub access key and secret key from logs.
   [GH-554]
 * builder/openstack: Properly scrub password from logs [GH-554]
+* builder/virtualbox: No panic if SSH host port min/max is the same. [GH-594]
+* builder/vmware: checks if `ifconfig` is in `/sbin` [GH-591]
+* builder/vmware: Host IP lookup works for non-C locales. [GH-592]
 * common/uuid: Use cryptographically secure PRNG when generating
   UUIDs. [GH-552]
 * communicator/ssh: File uploads that exceed the size of memory no longer